+++ src/clj/rems/application/model.clj
@@ -2,11 +2,12 @@
   (:require [clojure.set :as set]
             [clojure.test :refer [deftest is testing]]
             [com.rpl.specter :refer [ALL transform select]]
-            [medley.core :refer [distinct-by find-first map-vals update-existing update-existing-in]]
+            [medley.core :refer [dissoc-in distinct-by filter-vals find-first map-vals update-existing update-existing-in]]
             [rems.application.events :as events]
             [rems.application.master-workflow :as master-workflow]
             [rems.common.application-util :refer [applicant-and-members can-redact-attachment? is-applying-user?]]
             [rems.common.form :as form]
+            [rems.common.roles :refer [+handling-roles+]]
             [rems.common.util :refer [assoc-some-in conj-vec getx getx-in into-vec]]
             [rems.permissions :as permissions]
             [rems.json :as json]
@@ -51,8 +52,7 @@
   (-> application
       (assoc :application/modified (:event/time event))
       (assoc ::draft-answers (:application/field-values event))
-      (assoc-some-in [:application/duo :duo/codes] (when (:enable-duo rems.config/env)
-                                                     (:application/duo-codes event)))))
+      (assoc-some-in [:application/duo :duo/codes] (:application/duo-codes event))))
 
 (defmethod application-base-view :application.event/licenses-accepted
   [application event]
@@ -511,8 +511,8 @@
                                               :duo/restrictions (:restrictions dataset-code)}]
                nil)}))
 
-(defn- enrich-application-duo-matches [application]
-  (if-not (:enable-duo rems.config/env)
+(defn- enrich-application-duo-matches [application get-config]
+  (if-not (:enable-duo (get-config))
     application
     (let [duos (->> application :application/duo :duo/codes)
           matches (for [resource (:application/resources application)
@@ -789,6 +789,13 @@
        (transform [:application/attachments ALL] #(update % :attachment/user get-user))
        (transform [:application/attachments ALL] #(assoc % :attachment/redact-roles (allowed-redact-roles application %)))))
 
+(defn- enrich-workflow-anonymize-handling [application get-workflow]
+  (let [workflow-id (get-in application [:application/workflow :workflow/id])
+        workflow (get-workflow workflow-id)
+        anonymize-handling (get-in workflow [:workflow :anonymize-handling])]
+    (-> application
+        (assoc-some-in [:application/workflow :workflow/anonymize-handling] anonymize-handling))))
+
 (defn enrich-with-injections
   [application {:keys [blacklisted?
                        get-form-template
@@ -806,7 +813,7 @@
       enrich-field-visible ; uses enriched answers
       set-application-description ; uses enriched answers
       (update :application/resources enrich-resources get-catalogue-item)
-      enrich-application-duo-matches ; uses enriched resources
+      (enrich-application-duo-matches get-config) ; uses enriched resources
       (update-existing-in [:application/duo :duo/codes] duo/enrich-duo-codes)
       (enrich-workflow-licenses get-workflow)
       (update :application/licenses enrich-licenses get-license)
@@ -816,6 +823,7 @@
       (enrich-user-attributes get-user)
       (enrich-blacklist blacklisted?) ; uses enriched users
       (enrich-workflow-handlers get-workflow)
+      (enrich-workflow-anonymize-handling get-workflow)
       (enrich-deadline get-config)
       (enrich-super-users get-users-with-role)
       (enrich-workflow-disable-commands get-config get-workflow)
@@ -833,7 +841,7 @@
        (filterv #(= :visibility/public (:event/visibility %)))))
 
 (defn- censor-user [user]
-  (select-keys user [:userid :name :email :organizations :notification-email]))
+  (select-keys user [:userid :name :email]))
 
 (defn- censor-users-in-event [event]
   (-> event
@@ -857,13 +865,52 @@
       (update :application/events (partial mapv censor-users-in-event))
       (update :application/attachments (partial mapv #(update % :attachment/user censor-user)))))
 
+(def anonymous-handler {:userid "rems-handler"
+                        :name nil
+                        :email nil})
+
+(defn anonymize-users-in-attachments [attachments userids]
+  (vec
+   (for [attachment attachments
+         :let [userid (get-in attachment [:attachment/user :userid])]]
+     (if (some #{userid} userids)
+       (assoc attachment :attachment/user anonymous-handler)
+       attachment))))
+
+(defn anonymize-users-in-events [events userids]
+  (vec
+   (for [event events
+         :let [userid (:event/actor event)]]
+     (if (some #{userid} userids)
+       (-> event
+           (assoc :event/actor (:userid anonymous-handler))
+           (assoc :event/actor-attributes anonymous-handler))
+       event))))
+
+(defn get-handling-users [application]
+  (let [get-handling-roles #(clojure.set/intersection +handling-roles+ (set %))]
+    (set
+     (some->> (:application/user-roles application)
+              (filter-vals (comp seq get-handling-roles))
+              keys))))
+
+(defn apply-workflow-anonymization [application]
+  (if (get-in application [:application/workflow :workflow/anonymize-handling])
+    (let [handling-users (get-handling-users application)]
+      (-> application
+          (update :application/attachments anonymize-users-in-attachments handling-users)
+          (update :application/events anonymize-users-in-events handling-users)))
+    application))
+
 (defn- hide-sensitive-information [application]
   (-> application
       (dissoc :application/blacklist)
-      (update :application/events hide-sensitive-events) ; uses :event/public :event/visibility
+      (update :application/events hide-sensitive-events)
       (update :application/events (partial mapv #(dissoc % :event/public :event/visibility)))
-      (update :application/workflow dissoc :workflow.dynamic/handlers)
-      (update :application/workflow dissoc :workflow/voting)
+      apply-workflow-anonymization
+      (dissoc-in [:application/workflow :workflow.dynamic/handlers]
+                 [:application/workflow :workflow/voting]
+                 [:application/workflow :workflow/anonymize-handling])
       (dissoc :application/votes)
       hide-extra-user-attributes))