+++ src/clj/rems/service/test_data.clj
@@ -317,23 +317,24 @@
                                             :title "EGA workflow, a variant of default"
                                             :type :workflow/default
                                             :handlers handlers})
-        disabled-commands (test-helpers/create-workflow! {:actor owner
-                                                          :organization {:organization/id "nbn"}
-                                                          :title "Restricted workflow"
-                                                          :type :workflow/default
-                                                          :handlers handlers
-                                                          :licenses [link text]
-                                                          :disable-commands [{:command :application.command/invite-member}
-                                                                             {:command :application.command/close
-                                                                              :when/state [:application.state/returned]
-                                                                              :when/role [:applicant]}]})]
+        restricted (test-helpers/create-workflow! {:actor owner
+                                                   :organization {:organization/id "nbn"}
+                                                   :title "Restricted workflow"
+                                                   :type :workflow/default
+                                                   :handlers handlers
+                                                   :licenses [link text]
+                                                   :disable-commands [{:command :application.command/invite-member}
+                                                                      {:command :application.command/close
+                                                                       :when/state [:application.state/returned]
+                                                                       :when/role [:applicant]}]
+                                                   :anonymize-handling true})]
     {:default default
      :ega ega
      :decider decider
      :master master
      :auto-approve auto-approve
      :organization-owner organization-owner
-     :disabled-commands disabled-commands}))
+     :restricted restricted}))
 
 (defn- create-bona-fide-catalogue-item! [users]
   (let [owner (:owner users)
@@ -384,7 +385,9 @@
 (defn- create-applications! [catid catid-restricted users]
   (let [applicant (users :applicant1)
         approver (users :approver1)
-        reviewer (users :reviewer)]
+        approver2 (users :approver2)
+        reviewer (users :reviewer)
+        decider (users :decider)]
 
     (test-helpers/create-draft! applicant [catid] "draft application")
 
@@ -431,10 +434,44 @@
       (test-helpers/command! {:type :application.command/submit
                               :application-id app-id
                               :actor applicant})
-      (test-helpers/command! {:type :application.command/return
+      (test-helpers/command! {:type :application.command/remark
                               :application-id app-id
                               :actor approver
-                              :comment "Need more details"}))
+                              :comment "processing is due now"
+                              :public true})
+      (test-helpers/command! {:type :application.command/request-review
+                              :application-id app-id
+                              :actor approver2
+                              :reviewers [reviewer]
+                              :comment "please have a look"})
+      (test-helpers/command! {:type :application.command/remark
+                              :application-id app-id
+                              :actor reviewer
+                              :comment "application is missing key purpose"
+                              :public true})
+      (test-helpers/command! {:type :application.command/review
+                              :application-id app-id
+                              :actor reviewer
+                              :comment "not looking good in my opinion"})
+      (test-helpers/command! {:type :application.command/request-decision
+                              :application-id app-id
+                              :actor approver2
+                              :comment "please decide"
+                              :deciders [decider]})
+      (test-helpers/command! {:type :application.command/remark
+                              :application-id app-id
+                              :actor decider
+                              :comment "i agree with previous remarker"
+                              :public true})
+      (test-helpers/command! {:type :application.command/decide
+                              :application-id app-id
+                              :actor decider
+                              :decision :rejected
+                              :comment "unacceptable in current state"})
+      (test-helpers/command! {:type :application.command/return
+                              :application-id app-id
+                              :actor approver2
+                              :comment "need more details"}))
 
     (let [app-id (test-helpers/create-draft! applicant [catid] "approved & closed")]
       (test-helpers/command! {:type :application.command/submit
@@ -841,7 +878,7 @@
           catid (test-helpers/create-catalogue-item! cat)
           catid-restricted (test-helpers/create-catalogue-item!
                             (assoc cat
-                                   :workflow-id (:disabled-commands workflows)
+                                   :workflow-id (:restricted workflows)
                                    :title {:en "Default workflow (restricted)"
                                            :fi "Oletustyövuo (rajoitettu)"
                                            :sv "Standard arbetsflöde (begränsad)"}))]